Upgrading react-styleguidist-visual to latest
To pull in the new feature to allow a custom preview selector.

Using the custom preview selector to take a screenshot of the modal
after the click button is pressed in the example.

Updating the ModalDialog component to use onDismiss instead of
onDismissed. Using onDismissed is too late and the dialog gets into a
loop of being dismissed but popping back up.