fix(a11y): use provided labelId in custom ChoiceGroup label
Fabric's ChoiceGroupOption is passed a labelId that it expects
the label to use around the content. We weren't aware of or using this,
so ended up with an overall ChoiceGroupOption that had an aria-label
property on the INPUT field that didn't end up pointing to any valid
element.

This change picks up and uses the labelId to fix this.